archive/server/pyproject.toml
[tool.poetry]
name = "continuedev"
version = "0.1.104"
description = ""
authors = ["Nate Sesti <sestinj@gmail.com>"]
readme = "README.md"
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8.1"
boltons = "^23.0.0"
pydantic = "^1.10.8"
uvicorn = "^0.21.1"
python-dotenv = "^1.0.0"
nest-asyncio = "^1.5.6"
websockets = "^11.0.2"
urllib3 = "1.26.15"
posthog = "^3.0.1"
tiktoken = "^0.4.0"
jsonref = "^1.1.0"
jsonschema = "^4.17.3"
directory-tree = "^0.0.3.1"
anthropic = "^0.3.4"
chevron = "^0.14.0"
psutil = "^5.9.5"
pygithub = "^1.59.0"
meilisearch-python-async = "^1.4.8"
socksio = "^1.0.0"
ripgrepy = "^2.0.0"
bs4 = "^0.0.1"
replicate = "^0.11.0"
redbaron = "^0.9.2"
python-lsp-server = "^1.7.4"
huggingface-hub = "^0.16.4"
typer = "^0.9.0"
chromadb = "^0.4.12"
fastapi = "0.95.2"
tree-sitter = "^0.20.2"
tree-sitter-languages = "^1.7.0"
rich = "^13.6.0"
python-socketio = "^5.10.0"
pathspec = "^0.11.2"
tqdm = "^4.66.1"
html2text = "^2020.1.16"
pygit2 = "^1.13.2"
maturin = "^1.3.1"
openai = "^1.3.4"
aiohttp = "^3.9.0"
litellm = "^1.3.3"

[tool.poetry.scripts]
typegen = "continuedev.models.generate_json_schema:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.1"
pytest-asyncio = "^0.21.1"
ipykernel = "^6.26.0"

[tool.poetry.group.lint-ruff]
optional = true
[tool.poetry.group.lint-ruff.dependencies]
ruff = "==0.1.8"

[tool.poetry.group.lint-typing]
optional = true
[tool.poetry.group.lint-typing.dependencies]
mypy = "==1.7.1"
types-tqdm = "^4.66.0.5"
types-beautifulsoup4 = "^4.12.0.7"
types-chevron = "^0.14.2.5"
types-tree-sitter-languages = "^1.8.0.0"
types-aiofiles = "^23.2.0.0"
types-psutil = "^5.9.5.17"

[tool.mypy]
strict = true
show_error_codes = true
files = "continuedev,tests"
no_implicit_reexport = true

# Mypy 'whitelist'

[[tool.mypy.overrides]]
module = [
    "continuedev.*",
    "tests.*",
]
strict = false
disallow_untyped_calls = false
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_any_generics = false
disallow_subclassing_any = false
warn_return_any = false
strict_optional = false
disable_error_code = [
    "abstract",
    "arg-type",
    "assignment",
    "attr-defined",
    "call-arg",
    "import-not-found",
    "index",
    "list-item",
    "misc",
    "name-defined",
    "no-redef",
    "operator",
    "override",
    "return-value",
    "return",
    "syntax",
    "unused-coroutine",
    "union-attr",
    "var-annotated",
]

[[tool.mypy.overrides]]
module = [
    "boltons.*",
    "directory_tree.*",
    "huggingface_hub.*",
    "nest_asyncio.*",
    "litellm.*",
    "pygit2.*",
    "redbaron.*",
    "replicate.*",
    "requests.*",
    "ripgrepy.*",
    "socketio.*",
    "tiktoken_ext.*",
]
ignore_missing_imports = true


[build-system]
requires = ["poetry-core", "maturin>=1,<2"]
# build-backend = "poetry.core.masonry.api"
build-backend = "maturin"

[project]
name = "continuedev"
version = "0.1.90"
description = ""
authors = [{ name= "Nate Sesti", email = "<sestinj@gmail.com>" }]
readme = "README.md"
requires-python = ">=3.8.1"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]